+++ CHANGELOG.md
@@ -3,15 +3,16 @@
 ## 0.3.0 - Unreleased
 ### Added
 - Added conflcits resolver system
+- Added autoupdates to the PHAR via the `imm.phar self-update` command
 - Added better logic to guess the paths to the game and mods
 - Added ability to install/uninstall mods by text in addition to Steam ID (eg. `chud` will match `Chud Lite+`)
 - Added automatic resources extraction on Unix platforms
 
 ### Changed
 - The `mods:install` command now tries to install all mods, pass `-g` to only install non-LUA mods
-- Better Mac support
 
 ### Fixed
+- Better Mac support
 - Fixed binary always showing "dev version" even on stable builds
 
 ## 0.2.3 - 2017-03-09